refactor: MapSetting
