+++ src/components/editor/sidebar-element-settings-tab/index.tsx
@@ -3,38 +3,16 @@
 import { useEditor } from "~/components/editor/provider";
 import AdvanceSetting from "~/components/editor/sidebar-element-settings-tab/advance-setting";
 import LayoutSetting from "~/components/editor/sidebar-element-settings-tab/layout-setting";
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from "~/components/ui/accordion";
-import { Input } from "~/components/ui/input";
+import MapSetting from "~/components/editor/sidebar-element-settings-tab/map-setting";
 import { SheetHeader, SheetTitle } from "~/components/ui/sheet";
 
 type Props = {};
 
 export default function SidebarElementSettingsTab(props: Props) {
-  const { editor, dispatch } = useEditor();
-
-  const handleAddressInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    dispatch({
-      type: "UPDATE_ELEMENT",
-      payload: {
-        elementDetails: {
-          ...editor.state.selectedElement,
-          type: "map",
-          content: {
-            ...editor.state.selectedElement.content,
-            address: e.target.value,
-          },
-        },
-      },
-    });
-  };
+  const { editor } = useEditor();
 
   return (
-    <Accordion type="multiple" className="w-full" defaultValue={["Map"]}>
+    <div className="w-full">
       <SheetHeader className="border-b p-6">
         <SheetTitle>{editor.state.selectedElement.name} 설정</SheetTitle>
       </SheetHeader>
@@ -42,17 +20,7 @@ export default function SidebarElementSettingsTab(props: Props) {
       {!Array.isArray(editor.state.selectedElement.content) &&
         editor.state.selectedElement.type === "map" && (
           <>
-            <AccordionItem value="Map" className="border-0 border-t py-2">
-              <AccordionTrigger className="px-6 py-2">주소</AccordionTrigger>
-              <AccordionContent className="flex flex-col justify-center gap-2 px-6 py-1">
-                <Input
-                  id="address"
-                  placeholder="Enter your address"
-                  onChange={handleAddressInputChange}
-                  value={editor.state.selectedElement.content.address}
-                />
-              </AccordionContent>
-            </AccordionItem>
+            <MapSetting element={editor.state.selectedElement} />
           </>
         )}
 
@@ -63,6 +31,6 @@ export default function SidebarElementSettingsTab(props: Props) {
       )}
 
       <AdvanceSetting />
-    </Accordion>
+    </div>
   );
 }
